requirements.txt
Django==2.2
psycopg2-binary>=2.8
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
djangorestframework==3.9.2
djangorestframework-simplejwt==5.1.0
idna==3.3
PyJWT==2.3.0
pytz==2021.3
setuptools==57.0.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
wheel==0.36.2

